Настраивает сборку
