+++ public/index.html
@@ -45,6 +45,6 @@
       <!-- Контент -->
     </main>
 
-    <script src="../src/main.js"></script>
+    <script src="./bundle.js"></script>
   </body>
 </html>
