+++ src/components/siteBoardTasks.js
@@ -1,0 +1,7 @@
+const siteBoardTasks = () => {
+  return (
+    `<div class="board__tasks"></div>`
+  );
+};
+
+export {siteBoardTasks};
