+++ src/components/siteBoardTemplate.js
@@ -1,0 +1,7 @@
+const siteBoardTemplate = () => {
+  return (
+    `<section class="board container"></section>`
+  );
+};
+
+export {siteBoardTemplate};
